update: demo app
this commit adds a /whoami endpoint (json-API) in the main app to show the logged
in status of the user.

in addition we introduce a demo app which shows a third party can request the user
status (signed in, or signed out). this demo app is mounted into the same app to
avoid dealing with multiple projects (and ports) for the time being. the important
part is that the demo app does not include better-auth and could be written in any
other language.