update: implement passkey
this commits implements sign-in with passkey.

in order to be able to sign in with a passkey, the user must
first register and then add a passkey to their account (on the
profile page).

the profile page also lists passkeys (when there are any) and
allows for deletion of the passkey again.

in order to enable sign-up with passkey, we'd have to figure out
a flow where the user's email address is confirmed as well, since
better-auth wants to save emails along with each password.

not confirming the email would likely allow for weird account
takeover scenarios.