feat: redirect_url
this is the first implementation of the redirect_url feature in the
app.

this commit adds the redirect_url (from a url param) to the various
forms and handlers to pass it through login and registration flows
the user encounters.

for the configuration, I added a new key to the app config to support
URLs (maybe pattern matching later), to use this configuration in
the flows, we add a utils folder with helpers to validate the user
input.

while working on this code, I extract the forwardCookies method and
moved it to utils as well and changed its signature to use headers
directly.

I also refactored how I used the better-auth api calls for the
flows, mostly switched to setting the appropriate URL parameters and
switched to only returning headers, based on witch we can set cookies
and perform the expected redirects.

lbnl, I moved the code from the logout handler in (auth.js), to a new
handlers/logout.js so I could have POST/GET for /logout and it would
share the code. not entirely sure yet about the structure of this but
since hono doesn't give us anything, this might do?