chore(passkeys): refactor
i removed the passkeys code from the layout file and included it
via a "static" file which resides in static.

then I refactored the code the javascript handlers to address the
comments from the review, primarily to make it easier to follow in
regards to where and how the redirect_url is appended to the
various other URLs.

lbnl, extended the eslint configuration to allow files in "static"
to have access to the browser context (e.g. document, window).