fix(setup): disable husky
I've renamed the "prepare" task (for husky) as this task is
otherwise run each time we run "npm ci". because husky is an
optional devDependency, it interfered in the build process where
we don't want development dependencies.

the issue is elevated by the fact that husky does not get
installed, but after "npm ci", the "prepare" task tries to run
it, thus failing the entire task.

this also fixes a weirdness I encountered on my machine, where I
did not want to use husky, but every time I ran "npm ci" it would
set it up again and inject itself into my workflow.