feat(deployment): small docker container
this commit adds a docker container that can run our application.

in addition to the application code, we are setting up litestream
with configuration and a new entrypoint script which is used to
initialise or restore the database when the container is started.

to unify the configuration (somewhat), we made the location of the
"auth.db" configurable using a new DATABASE_URL environment variable
which is used by the app and the litestream configuration